DataRef: defer creation of platform specific object until needed
This allows to use DataRef for passing information within C++ without
the limitations of the platform objects (e.g. wasm emscripten::val
cannot be moved between threads).
Furthermore, this allows using DataRef agnostic of the type of bindings
that will be generated for it (if any) -- the DATAREF_JNI/WASM/OBJC are
no longer necessary.

The optimized case for the initializing the DataRef from a (moved) string
is not re-implemented here out of sheer laziness. We dont use this anyway,
and its just more complexity for no gain.